refactor: adjust project structure
