+++ modules/common/make-path-absolute.js
