+++ modules/fs/move.js
@@ -1,5 +1,5 @@
 import { copy } from "./copy.js"
-import { remove } from "./delete.js";
+import { remove } from "./remove.js";
 
 export const move = async (srcPath, destDirPath) => {
   await copy(srcPath, destDirPath);
